general updates
Pass robot state text through functions as it's not right that tasks applied it.
Renamed Task.tasks to be Task.plannedTasks as we will add robot adhoc tasks for edge case fixing and other oddities.
Pathfinder accuracy changed to walk right to the spot to avoid needing to mine things at the end of paths.
Add simple handler for getting stuck when walking path.