Deconstruct Area - WIP 3
move from the sorted double chunk lists to objects with meta data about their lists contents. As on further development we wanted to start mid sorted list, but didn't know where we were in the list without adding even more cross linked indexes. And they were only beneficial on the occasional checking for next chunk.
Moved a lot of inline notes in to TODO list to try and manage it better.