Deconstruct Area - WIP 7
Deconstruction working for single and double robots.
Account for robots mining speed.
Fix bug in getting nearest as it was janky to watch on an irregular entity built order.
Let the robots plan more per tick as their ups usage at that stage was much lower than when they were mining and moving.