Utils StyleData change
remove the style version from the font name used in the lists. As we don't use it in the other style types.
This would be a breaking change, but this is the WIP version of Utility and has already made a breaking change version number change since the main Utils release.