EventScheduler Remove functions return count
Make EventScheduler.RemoveScheduledOnceEvents() and EventScheduler.RemoveScheduledEventFromEachTick() both return the count of the number of entries they removed. Avoids using a separate IsEventScheduledXXXXX() to know if something was actually going to be removed. Came up from Muppet Streamer and wanting to confirm to a user if they script remove a Delayed Lua function that doesn't exist.