Events.RemoveHandler() review
This was never used before and has a number of issues that meant before it breaks the wider data integrity of the events, breaks when a handler is added-removed-added, doesn't update the events wider filter list, doesn't cancel the core Factorio event registrations when the last handler is removed.
Also some tidyup and related changes to the event handler recording process.