BiterSelection - use Mod rather than global
When it used global it would be invalid/error if evolution is unchanged across a save/load and a biter mod has been removed during the gap. The data should be stored in Lua global and not Factorio global. We do this for prototype-attributes for this reason. As we don't need to persist the result for the current evolution, just cache it to avoid mass near simultaneous lookups.